+++ mcapdecode/mcapdecode/src/lib.rs
@@ -1,9 +1,7 @@
 #[cfg(feature = "arrow")]
-mod arrow_ext;
-#[cfg(feature = "arrow")]
-pub use arrow_ext::{McapReaderArrowExt, RecordBatchOptions};
-#[cfg(feature = "arrow")]
 pub use mcapdecode_arrow as arrow;
+#[cfg(feature = "arrow")]
+pub use mcapdecode_arrow::{McapReaderArrowExt, RecordBatchOptions};
 pub use mcapdecode_core as core;
 pub use mcapdecode_reader::{
     McapReader, McapReaderBuilder, McapReaderError, PreparedTopic, RawMessage, ReadOptions,